requirements.txt
numpy==1.23.5
tensorflow==2.9.1
fastapi>=0.85.0
uvicorn>=0.18.3
pillow>=9.0.0
python-multipart>=0.0.5
